[FE] feat: SideBar 구조 변경 및 주석 추가 (#32)
